Renamed sendLinks to links and added list functionality
